% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eps_AC_lm.R
\name{epsAC.lm}
\alias{epsAC.lm}
\title{Fit linear model to EPS-full data}
\usage{
epsAC.lm(formula, confounder)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}}, that
describes the linear regression model to be fitted, see details}

\item{confounder}{confounding (non-genetic) covariates}
}
\value{
Maximum likelihood estimates of the model parameters,
with 95 percent confidence intervals
\item{coefficients}{a vector of maximum likelihood estimates of the coefficients}
\item{ci}{a matrix of confidence intervals for the coefficients}
\item{sigma}{the estimated standard deviation}
}
\description{
Fit linear model to EPS-full data
}
\details{
The \code{\link[stats]{formula}} object is of the type
y~xe+xg, which describes a regression model, y=a+be*xe+bg*xg+e
assuming a normal distribution for the residuals (e). The covariate
xe is a non-genetic/environmental covariate (optional).
The covariate xg is a SNP (single-nucleotide polymorphism).
The variables are taken from the environment that the
function is called from.
Both xe and xg can be matrices.

The EPS all case design is such that the SNP genotype is only observed
for individuals with high and low values of the phenotype \code{y}.
For remaining individuals, the unobserved genotype most be coded as NA.

Confounders are discrete covariates (xe) and the distribution of xg is
modelled for each level of unique value of xe.
}
\examples{

}
