% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eps_AC_testrare.R
\name{epsAC.rv.test}
\alias{epsAC.rv.test}
\title{Test for associations under the EPS complete-case design}
\usage{
epsAC.rv.test(nullmodel, RV, confounder = FALSE, cx, method = "naive",
  weights)
}
\arguments{
\item{nullmodel}{an object of class \code{\link[stats]{formula}}, that
describes the linear regression model under the null hypothesis}

\item{RV}{a matrix of genetic variants to be tested against the null}
}
\value{
\code{epsAC.rv.test} returns for each genetic variant:
\item{statistic}{the score test statistic}
\item{p.value}{the P-value}
}
\description{
\code{epsAC.rv.test} performs a score test for common genetic variants
under the EPS complete-case design
}
\details{
nanana
}
\examples{
N = 5000 # Number of individuals in a population
xe1 = rnorm(n = N, mean = 2, sd = 1) # Environmental covariate
xe2 = rbinom(n = N, size = 1, prob = 0.3) # Environmental covariate
xg1 = sample(c(0,1,2),N,c(0.4,0.3,0.3), replace = TRUE) # SNP
xg2 = sample(c(0,1,2),N,c(0.5,0.3,0.2), replace = TRUE) # SNP
# Model parameters
a = 50; be1 = 5; be2 = 8; bg1 = 0.3; bg2 = 0.6; sigma = 2
# Generate response y
y = rnorm(N, mean = a + be1*xe1 + be2*xe2 + bg1*xg1 + bg2*xg2, sd = sigma)
# Identify extremes, here upper and lower 25\% of population
u = quantile(y,probs = 3/4,na.rm=TRUE)
l = quantile(y,probs = 1/4,na.rm=TRUE)
}
