% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eps_AC_testGE.R
\name{epsAC.test.GE}
\alias{epsAC.test.GE}
\title{Test for gene-environment associations under the EPS-AC design}
\usage{
epsAC.test.GE(nullmodel, G, E, confounder)
}
\arguments{
\item{nullmodel}{an object of class \code{\link[stats]{formula}}, that
describes the linear regression model under the null hypothesis
(including the genetic covariate)}

\item{G}{The genetic covariate in the GE interaction}

\item{E}{The environmental covariate in the GE interaction}

\item{confounder}{(optional) vector of names of confounding (non-genetic) covariates}
}
\value{
\code{epsAC.testGE} returns
\item{statistic}{thescore test statistic}
\item{p.value}{the P-value}
}
\description{
\code{epsAC.test.GE} performs a likelihood ratio test for gene-environment
interaction variables under the EPS-AC design
}
\details{
The \code{nullmodel} \code{\link[stats]{formula}} object is of the type
y~xe+xg, which describes a regression model, y=a+be*xe+bg*xg+e
assuming a normal distribution for the residuals (e). The covariate
xe is a non-genetic/environmental covariate (optional).
The covariate xg is one or more, genetic markers where missing values
are coded as NA. Missing-mechanism must be MCAR or MAR.

Confounders are discrete covariates (xe) and the distribution of xg is
modelled for each level of unique value of xe.

The test considers a regression model y=a+be*xe+bg*xg+b*xe*xg+e,
where b=0 under the null hypothesis. The output
of the function gives the test statistic and p-value for the test of
H0: b=0. The genetic covarite (G) and the environmental covariate (E) for
the interaction must be specified, G with missing values coded as NA and
E with no missing values. G must also be included in the null model.
}
\examples{


}
