% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eps_AC_test.R
\name{epsAC.test}
\alias{epsAC.test}
\title{Score test EPS-AC}
\usage{
epsAC.test(nullmodel, xg, confounder)
}
\arguments{
\item{nullmodel}{an object of class \code{\link[stats]{formula}}, that
describes the linear regression model under the null hypothesis, see details}

\item{xg}{a matrix of variables to be tested against the null (NA for
not genotyped individuals)}

\item{confounder}{confounding (non-genetic) covariates}
}
\value{
\code{epsAC.test} returns
\item{statistic}{the score test statistic}
\item{p.value}{the P-value}
}
\description{
\code{epsAC.test} performs a score test for common genetic variants
under the EPS all-case design
}
\details{
The \code{nullmodel} \code{\link[stats]{formula}} object is of the type
y~xe, which describes a regression model, y=a+be*xe+e
assuming a normal distribution for the residuals (e). The covariate
xe is a non-genetic/environmental covariate (optional).
The variables are taken from the environment that the
function is called from.
The null hypothesis bg=0 is tested for the model y=a+be*xe+bg*xg+e.
The covariate xg is one or more, genetic markers where missing values
are coded as NA. Missing-mechanism must be MCAR or MAR.
If xg is a matrix, each variant (column) is tested against the null model.

Confounders are discrete covariates (xe) and the distribution of xg is
modelled for each level of unique value of xe.
}
\examples{
# Testing
epsAC.test(phenoCV ~ EPSxe,xg = gCV)
epsAC.test(phenoCV ~ EPSxe1 + EPSxe2,xg = gCV,confounder = EPSxe1)
}
