% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eps_CC_lm.R
\name{epsCC.lm}
\alias{epsCC.lm}
\title{Fit linear model to EPS-only data}
\usage{
epsCC.lm(formula, l, u)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}}, that
describes the linear regression model to be fitted, see details}

\item{l}{cutoff for lower extreme, can be sample-specific or general}

\item{u}{cutoff for upper extreme, can be sample-specific or general}
}
\value{
Maximum likelihood estimates of the model parameters,
with 95 percent confidence intervals
\item{coefficients}{a vector of maximum likelihood estimates of
the coefficients}
\item{ci}{a matrix of confidence intervals for the coefficients}
\item{sigma}{the estimated standard deviation}
}
\description{
\code{epsCC.lm} fits a normal linear regression model to EPS-only data
}
\details{
The \code{\link[stats]{formula}} object is of the type
y~xe+xg, which describes a regression model, y=a+be*xe+bg*xg+e
assuming a normal distribution for the residuals (e). The covariate
xe is a non-genetic/environmental covariate (optional).
The covariate xg is a SNP (single-nucleotide polymorphism).
The variables are taken from the environment that the
function is called from.
Both xe and xg can be matrices.

The EPS-only design is such that data is only available
for individuals with high and low values of the phenotype \code{y}. The
cut-offs \code{l} and \code{u} that specify the sampling must be given

Thus, the data set must consist of observations only for extreme individuals
\code{y > u} or \code{y < l} where \code{y} is the response (phenotype)
of the linear regression model.
}
\examples{


}
