% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsonlytest.R
\name{epsonly.test}
\alias{epsonly.test}
\title{Test for associations under the EPS-only design}
\usage{
epsonly.test(nullmodel, SNP, cutoffs, randomindex, onebyone = TRUE)
}
\arguments{
\item{nullmodel}{an object of class \code{\link[stats]{formula}}, that
describes the linear regression model under the null hypothesis}

\item{SNP}{a matrix of variables to be tested against the null}

\item{cutoffs}{a vector \code{c(l,u)} of the lower and upper cut-offs used
for extreme sampling}

\item{randomindex}{a vector which indicates if observations are from a random
or extreme sample}

\item{onebyone}{\code{TRUE} if genetic variables should be tested one by one
for inclusion in the model, default \code{TRUE}}
}
\value{
\code{epsonly.test} returns
\item{statistic}{the value of the score test statistic}
\item{parameter}{the degrees of freedom of the statistic}
\item{p.value}{the P-value for the test}
}
\description{
\code{epsonly.test} performs a score test for genetic variables
in the EPS-only design
}
\details{
The \code{nullmodel} \code{\link[stats]{formula}} object is of the type
y~xe, which describes a regression model, y=a+be*xe+e
assuming a normal distribution for the residuals (e). The covariate
xe is a non-genetic/environmental covariate (optional).
The variables are taken from the environment that the
function is called from.
The test considers a regression model y=a+be*xe+bg*xg+e, where xg is a
genetic covariate, and where bg=0 under the null hypothesis. The output
of the function gives the test statistic and p-value for the test of
H0: bg=0.
The covariate \code{xg} is a SNP (single-nucleotide polymorphism).
Both xe and xg can be matrices.

The EPS-only design is such that data is only available
for individuals with high and low values of the phenotype \code{y}. The
cut-offs \code{l} and \code{u} that specify the sampling must be specified
in the \code{cutoffs} argument.

Thus, the data set must consist of observations only for extreme individuals
\code{y > u} or \code{y < l} where \code{y} is the response (phenotype)
of the linear regression model.
}
\examples{
N = 5000 # Number of individuals in a population
xe1 = rnorm(n = N, mean = 2, sd = 1) # Environmental covariate
xe2 = rbinom(n = N, size = 1, prob = 0.3) # Environmental covariate
xg1 = sample(c(0,1,2),N,c(0.4,0.3,0.3), replace = TRUE) # SNP
xg2 = sample(c(0,1,2),N,c(0.5,0.3,0.2), replace = TRUE) # SNP
# Model parameters
a = 50; be1 = 5; be2 = 8; bg1 = 0.3; bg2 = 0.6; sigma = 2
# Generate response y
y = rnorm(N, mean = a + be1*xe1 + be2*xe2 + bg1*xg1 + bg2*xg2, sd = sigma)
# Identify extremes, here upper and lower 25\% of population
u = quantile(y,probs = 3/4,na.rm=TRUE)
l = quantile(y,probs = 1/4,na.rm=TRUE)
extreme = (y < l) | (y >= u)
# Create the EPS-only data set
y = y[extreme]
xe1 = xe1[extreme]
xe2 = xe2[extreme]
xg1 = xg1[extreme]
xg2 = xg2[extreme]
xg = as.matrix(cbind(xg1,xg2))
xe = as.matrix(cbind(xe1,xe2))

epsonly.test(y~xe,SNP=xg,cutoffs = c(l,u))
epsonly.test(y~xe,SNP=xg,cutoffs = c(l,u),onebyone = FALSE)
}
