% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eps_AC_testrare.R
\name{epsAC.rv.test}
\alias{epsAC.rv.test}
\title{Score test EPS-AC rare variants}
\usage{
epsAC.rv.test(nullmodel, xg, confounder, method = "simple", weights)
}
\arguments{
\item{nullmodel}{an object of class \code{\link[stats]{formula}}, that
describes the linear regression model under the null hypothesis}

\item{xg}{a matrix of variables to be tested against the null (NA for
not genotyped individuals)}

\item{confounder}{(optional) vector of names of confounding (non-genetic) covariates}

\item{method}{testing the burden using \code{simple}, \code{collapse} or
\code{varcomp} method, see details}

\item{weights}{optional weights}
}
\value{
\code{epsAC.rv.test} returns for the whole burden of variants:
\item{statistic}{the score test statistic}
\item{p.value}{the P-value}
}
\description{
\code{epsAC.rv.test} performs a score test for a burden of
rare genetic variants under the EPS all-case design
}
\details{
The \code{nullmodel} \code{\link[stats]{formula}} object is of the type
y~xe, which describes a regression model, y=a+be*xe+e
assuming a normal distribution for the residuals (e). The covariate
xe is a non-genetic/environmental covariate (optional).
The variables are taken from the environment that the
function is called from.
The null hypothesis bg=0 is tested for the model y=a+be*xe+bg*xg+e.
The covariate \code{xg} is a burden of several rare genetic variants,
where missing values are coded as NA. Missing-mechanism must be MCAR or MAR.
Missing-mechanism assumed equal for all genetic variants.

Confounders are discrete covariates (xe) and the distribution of xg is
modelled for each level of unique value of xe.

The \code{simple} method uses a standard score test to test the burden,
the \code{collapse} method tests the (weighted) sum of all variants in the
burden, while the \code{varcomp} method is a (weighted) variance
component score test.

The \code{varcomp} method is a special case of the popular SKAT method
for a linear weighted kernel under extreme phenotype sampling. The p-value is found
using the function \code{\link[CompQuadForm]{davies}} in the CompQuadForm
package.
}
\examples{
# Simple test
epsAC.rv.test(phenoRV~EPSxe,xg=gRV)
# Collapsing test
epsAC.rv.test(phenoRV~EPSxe,xg=gRV,method = "collapse")
# Variance component test
epsAC.rv.test(phenoRV~EPSxe,xg=gRV,method = "varcomp")

Control for include confounders
epsAC.rv.test(phenoRV~EPSxe1 + EPSxe2,xg=gRV,confounder = EPSxe1)
}
\references{
\insertRef{quadRpackage}{extremesampling},

\insertRef{wu2011SKAT}{extremesampling}
}
\seealso{
\code{\link[SKAT]{SKAT}} for the SKAT test for random samples,
\code{\link[CompQuadForm]{davies}} for the Davies method,
\code{\link{epsAC.test}} for a common variant SNP by SNP test for
all-case extreme sampling
}
