% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eps_CC_test.R
\name{epsCC.test}
\alias{epsCC.test}
\title{Score test EPS-CC}
\usage{
epsCC.test(nullmodel, xg, l, u, randomindex)
}
\arguments{
\item{nullmodel}{an object of class \code{\link[stats]{formula}}, that
describes the linear regression model under the null hypothesis}

\item{xg}{a matrix of genetic variants to be tested against the null}

\item{l}{cutoff for lower extreme, can be sample-specific or general}

\item{u}{cutoff for upper extreme, can be sample-specific or general}

\item{randomindex}{a binary vector that indicates if samples are random
or extreme}
}
\value{
\code{epsCC.test} returns for each genetic variant:
\item{statistic}{the score test statistic}
\item{p.value}{the P-value}
}
\description{
\code{epsCC.test} performs a score test for common genetic variants
under the EPS complete-case design
}
\details{
The \code{nullmodel} \code{\link[stats]{formula}} object is of the type
y~xe, which describes a regression model, y=a+be*xe+e
assuming a normal distribution for the residuals (e). The covariate
xe is a non-genetic/environmental covariate (optional).
The variables are taken from the environment that the
function is called from.
The null hypothesis bg=0 is tested for the model y=a+be*xe+bg*xg+e.
The covariate xg is one or more genetic markers.

Variables are only available
for individuals with high and low values of the phenotype \code{y};
(\code{y < l} or \code{y > u}), and potentialy some randomly sampled
individuals. The cut-offs \code{l} and \code{u} that specify the
sampling must be given in the \code{cutoffs} argument.
}
\examples{
epsCC.test(phenoCV_CC ~ xeCV_CC,xg = gCV_CC,lCV,uCV)
}
